% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.utilities.R
\name{plot.brier.auc}
\alias{plot.brier.auc}
\title{Plot Brier Score and time-dependent AUC.}
\usage{\method{plot}{brier.auc}(o, loo = 25, show.plots = TRUE)}
\arguments{
\item{x}{rfsrc object (survival).}
\item{loo}{Number of monte carlo sets used for the leave-one-out (loo)
  bootstrap used for Brier and time dependent AUC standard errors. Only
  applies to survival.}
\item{show.plots}{Should plots be displayed?}
}
\description{
Brier score and time-dependent AUC for a rfsrc (survival) object.
}
\examples{
## survival
data(pbc, package = "randomForestSRC")
o <- rfsrc(Surv(days, status) ~ ., data = pbc)
plot.brier.auc(o)

## same example, but using show.plots option
g <- plot.brier.auc(o, show.plots = FALSE)
plot(g)
}
